+++ Beyond.Todo.Application/Commands/RegisterProgressionCommand.cs
@@ -2,4 +2,4 @@
 
 namespace Beyond.Todo.Application.Commands;
 
-public sealed record RegisterProgressionCommand(int Id, DateTime Date, decimal Percent) : IRequest<Unit>;
+public sealed record RegisterProgressionCommand(int Id, DateTime Date, int Percent) : IRequest<Unit>;
